add the ability to start and stop device polling
